[chore] #320 Observer 싱글톤 클래스 파일 분리
