+++ POME/Presentation/Utils/RecordObserver.swift
@@ -1,0 +1,19 @@
+//
+//  RecordObserver.swift
+//  POME
+//
+//  Created by 박소윤 on 2023/05/18.
+//
+
+import Foundation
+import RxSwift
+
+class RecordObserver{
+    
+    static let shared = RecordObserver()
+    
+    private init(){ }
+    
+    let generateRecord = PublishSubject<Void>()
+    let registerSecondEmotion = PublishSubject<Void>()
+}
